Added Service Chaining models compatible with Neutron SFC API.
Service - a network function accessible via a port. It may be either L2, trans-
parent to traffic, or L3 in which case it only processes packets with its own
MAC in the Ethernet destination, and re-transmits them with its own MAC in the
Ethernet source and its next hop gateway's MAC in the ethernet destination.

Service Chain - a sequence of network functions through which packets should be
directed. The Chain has a name, a port at which it should be inserted, and a
set of traffic classifiers specifying which traffic should be directed through
the chain.

Service Chain Element - an element specified that a specific service is part of
a specific chain, the vlan that should be added to traffic in order to identify
the element/chain (if the service re-emits the redirected traffic), the order
in which the service should be applied if there are others in the chain, the
failure model, and whether the service should be treated as routed or L2 bump-
in-the-wire. If two or more elements in the same chain have the same position,
then traffic should be load-balanced among them.

Change-Id: I8333034d81fa136feeebc5175f78f2b1c81f58c5
Signed-off-by: Giuseppe (Pino) de Candia <gdecandia@midokura.com>